backend/requirements.txt
django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
dj-database-url==2.2.0
whitenoise==6.8.2
gunicorn==23.0.0
psycopg[binary]==3.2.13
